refactor(Piece): 같은 팀인지 확인하는 메서드 긍정형으로 변경
